Key tree property tests
Key tree module is a candidate to use property tests on as it mostly deals with
manipulating a single data structure and functions are referentially
transparent, that is, they aren't many side-effects like IO generated.

Property tests are currently run as part of the EUnit testing framework. In
the future it might be interesting to have a longer or more thourough runs
to perhaps catch more edge cases.

To run the tests used Triq. It is an Apache licensed property testing tool. It
is not as full features as the commerical and the other, GPL-licensed one, but
seems to have the basics and should be easy to integrate with CouchDB's code.

To run the test:

make eunit apps=couch suites=couch_key_tree_prop_tests

To play with the code in the interpreter can try this script:

```
set +e
make > /dev/null
exec erl \
 -pa src/couch/ebin \
 -pa src/triq/ebin/ \
 -pa src/config/ebin \
 -eval "application:start(config), code:load_file(couch_key_tree), code:load_file(triq)"
```